refactor: replace delete local data modal with a reusable component
